+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- Fix `onet_api_url` to depend on the connected chain and remove default endpoint.
+
 ## [0.10.1] - 2023-07-14
 
 ## New